Refactoring: compare to class diagram.
