+++ patterns/state/manipulator_state/app/base_manipulation.dart
