+++ patterns/state/manipulator_state/states/_/sub_states/parent_state.dart
@@ -4,7 +4,7 @@ import '../../selections/selection_state.dart';
 import 'child_state.dart';
 import '../../../shapes/shape.dart';
 
-class ParentState<T extends Shape> extends SelectionState<T> {
+abstract class ParentState<T extends Shape> extends SelectionState<T> {
   ParentState({
     required super.selectedShape,
   });
