+++ patterns/state/manipulator_state/states/selections/resizable_markers/top_right_marker_state.dart
@@ -1,4 +1,4 @@
-import '../../_/corner_marker.dart';
+import '../../_/marker.dart';
 
 class TopRightMarkerState extends Marker {
   TopRightMarkerState({
